[Sync] Add Sampling and Method change check during sync, also CruiseSyncer refactor
