+++ src/CruiseDAL.V3.Sync.Test.Spec/CruiseDAL.V3.Sync.Test.Spec.csproj
@@ -34,4 +34,6 @@
 
 	<Import Project="..\CruiseDAL.TestCommon\CruiseDAL.TestCommon.projitems" Label="Shared" />
 
+	<Import Project="..\CruiseDAL.TestCommon.V3\CruiseDAL.TestCommon.V3.projitems" Label="Shared" />
+
 </Project>
