+++ src/CruiseDAL.V3.Sync/Syncers/StratumTemplateSyncer.cs
@@ -1,0 +1,50 @@
+﻿using CruiseDAL.V3.Models;
+using FMSC.ORM.Core;
+using System.Data.Common;
+using System.Linq;
+
+namespace CruiseDAL.V3.Sync.Syncers
+{
+    public class StratumTemplateSyncer : TableSyncerBase
+    {
+        public StratumTemplateSyncer() : base(nameof(StratumTemplate))
+        {
+        }
+
+        public override TableSyncResult SyncRecords(string cruiseID, DbConnection source, DbConnection destination, TableSyncOptions options, IExceptionProcessor exceptionProcessor)
+        {
+            var syncResult = new TableSyncResult(nameof(StratumTemplate));
+
+            var flags = options.Template;
+            if (flags == SyncOption.Lock) { return syncResult; }
+
+            var where = "CruiseID = @CruiseID AND StratumTemplateName = @StratumTemplateName";
+            var sourceItems = source.From<StratumTemplate>()
+                .Where("CruiseID = @p1").Query(cruiseID);
+
+            foreach (var i in sourceItems)
+            {
+                var match = destination.From<StratumTemplate>().Where(where).Query2(i).FirstOrDefault();
+
+                if (match == null)
+                {
+                    var hasTombstone = destination.From<StratumTemplate_Tombstone>()
+                                                    .Where(where).Count2(i) > 0;
+
+                    if (flags.HasFlag(SyncOption.ForceInsert)
+                                || (hasTombstone == false && flags.HasFlag(SyncOption.Insert)))
+                    {
+                        destination.Insert(i, persistKeyvalue: false, exceptionProcessor: exceptionProcessor);
+                        syncResult.IncrementInserts();
+                    }
+                }
+                else if (ShouldUpdate(i.Modified_TS, match.Modified_TS, flags))
+                {
+                    destination.Update(i, whereExpression: where, exceptionProcessor: exceptionProcessor);
+                    syncResult.IncrementUpdates();
+                }
+            }
+            return syncResult;
+        }
+    }
+}
